Fix compile error on syn 1.0.58
imgui-inspect-derive was depending on the private API details of `syn`, the re-export of `quote` types, which in https://github.com/dtolnay/syn/commit/957840eba2c7f11c95e0227760d78dc481a91de7 was renamed and further deprecated/discouraged.

This rename caused build errors of the imgui-inspect-derive crate when using `syn` 1.0.58 patch versions.

Simply switched here to use the type directly from `quote` instead of relying on the private exports of `syn`.

This is fairly important to get in ASAP as this crate breaks when using newer patch versions of `syn` and `serde`